+++ lib/ReactViews/Map/Panels/SharePanel/Print/PrintWorkbench.tsx
@@ -1,7 +1,9 @@
 import React from "react";
 import Workbench from "../../../../../Models/Workbench";
 
-import CatalogMemeberMixin from "../../../../../ModelMixins/CatalogMemberMixin";
+import CatalogMemeberMixin, {
+  getName
+} from "../../../../../ModelMixins/CatalogMemberMixin";
 import MappableMixin from "../../../../../ModelMixins/MappableMixin";
 import { BaseModel } from "../../../../../Models/Definition/Model";
 import DiscretelyTimeVaryingMixin from "../../../../../ModelMixins/DiscretelyTimeVaryingMixin";
@@ -11,14 +13,13 @@ import SelectableDimensions, {
   filterSelectableDimensions,
   findSelectedValueName
 } from "../../../../../Models/SelectableDimensions";
-import { useTranslation } from "react-i18next";
+import CatalogMemberMixin from "../../../../../ModelMixins/CatalogMemberMixin";
 
 interface Props {
   workbench: Workbench;
 }
 
-const renderDisplayVariables = (catalogItem: BaseModel & any) => {
-  const { t } = useTranslation();
+const renderDisplayVariables = (catalogItem: BaseModel) => {
   if (SelectableDimensions.is(catalogItem)) {
     return filterSelectableDimensions(DEFAULT_PLACEMENT)(
       catalogItem.selectableDimensions
@@ -38,25 +39,27 @@ const renderLegend = (
     BaseModel &
     any
 ) => {
-  if (!catalogItem.isMappable) {
+  if (!MappableMixin.isMixedInto(catalogItem)) {
     return null;
   }
 
   return (
     <div key={catalogItem.uniqueId} className="layer-legends">
-      <div className="layer-title">{catalogItem.name}</div>
-      {catalogItem.currentTime && (
+      <div className="layer-title">{getName(catalogItem)}</div>
+      {DiscretelyTimeVaryingMixin.isMixedInto(catalogItem) && (
         <div className="layer-time">Time: {catalogItem.currentTime}</div>
       )}
-      <Legend forPrint={true} item={catalogItem} />
+      {CatalogMemberMixin.isMixedInto(catalogItem) && (
+        <Legend forPrint={true} item={catalogItem} />
+      )}
     </div>
   );
 };
 
-const WorkbenchItem = ({ item }: { item: BaseModel | any }) => {
+const WorkbenchItem = ({ item }: { item: BaseModel }) => {
   return (
     <div className="WorkbenchItem">
-      <h3>{item.name}</h3>
+      <h3>{getName(item)}</h3>
       {renderDisplayVariables(item)}
       <div>{renderLegend(item)}</div>
     </div>